tests/wpt/tests/clipboard-apis/async-unsanitized-standard-html-formats-write-read.tentative.https.html

<!doctype html>
<meta charset="utf-8">
<title>Async Clipboard unsanitized HTML write -> Async Clipboard unsanitized HTML read test</title>
<link rel="help" href="https://w3c.github.io/clipboard-apis/#async-clipboard-api">
<body>Body needed for test_driver.click()</body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/user-activation.js"></script>
<script>
'use strict';

// This function removes extra spaces between tags in html. For example, the
// following html: "<p> Hello </p>   <body> World </body>" would turn into this
//           html: "<p> Hello </p> <body> World </body>"
// We remove the extra spaces because in html they are considered equivalent,
// but when we are comparing for equality the spaces make a difference.
function reformatHtml(html) {
  const parser = new DOMParser();
  const htmlString =
        parser.parseFromString(html, 'text/html').documentElement.innerHTML;
  const reformattedString = htmlString.replace(/\>\s*\</g, '> <').replace(/<!--[a-zA-Z]*?-->/g, '');
  return reformattedString;
}

// Writes a payload with HTML content and checks to ensure the correct data
// was written successfully.
promise_test(async t => {
  await test_driver.set_permission({name: 'clipboard-read'}, 'granted');
  await test_driver.set_permission({name: 'clipboard-write'}, 'granted');

  // Create and write unsanitized version of standard HTML format.
  const format1 = 'text/html';
  // The string must be concatenated in this way because the html parser
  // will recognize a script tag even in quotes as a real script tag. By
  // splitting it up in this way we avoid that error.
  const html_script = '<script>const a = 5;</scr'
                          + 'ipt> <p>Hello World</p>';
  const textInput = '<head><style>color:blue</style><head>' +
                    '<body><p>Hello</p>' + html_script + '</body>';
  const blobInput1 = new Blob([textInput], {type: format1});
  const clipboardItemInput = new ClipboardItem({[format1]: blobInput1});
  await waitForUserActivation();
  await navigator.clipboard.write([clipboardItemInput]);

  // Read unsanitized version of HTML format.
  await waitForUserActivation();
  const clipboardItems =
    await navigator.clipboard.read({unsanitized: ["text/html"]});
  assert_equals(clipboardItems.length, 1);
  const clipboardItem = clipboardItems[0];
  assert_true(clipboardItem instanceof ClipboardItem);

  const blobOutput1 = await clipboardItem.getType(format1);
  assert_equals(blobOutput1.type, format1);

  const data1 = await (new Response(blobOutput1)).text();
  const outputHtml = reformatHtml(data1);
  const expectedHtml = '<html>' + textInput + '</html>';
  const unsanitizedExpectedHtml = reformatHtml(expectedHtml);
  assert_equals(outputHtml, unsanitizedExpectedHtml);
}, 'Verify write and read unsanitized content to the clipboard given text/html format as input');
</script>
